Removed transitions between mesoscopic subvolumes that are not the same type
- initial placeholder implementation of transitions between mesoscopic
subvolumes that are of different types (e.g., normal and surface).
Normal diffusion between such regions will not be possible, but
reactions with at a surface should be represented as a transitions
to/from neighboring normal subvolumes. For now, transition rate is set
to 0. The structure of chemical reactions may need to be expanded to
accommodate this, particularly for non-trivial surface reactions.
- presence of a membrane surface between 2 subvolumes is not considered
when determining transition rate between them.