Updated actor intersection detection for 2D regions and surfaces
- added dimension and effectiveDim members to region structure to make
it easier when checking how regions need to be treated
- updated the "volume" calculation for the intersection of a region and
a boundary. Will return area if region is effectively 2D.
- imposed that actor boundary cannot intersect a surface whose dimension
is not the same as its effective dimension (e.g., a 3D surface that is a
3D shape is effectively just 2D)
- improved test for whether region is fully inside an actor by comparing
boundaries and not relying on volume calculations
- updated mesoscopic subvolume search range to account for 2D regions
and surfaces. Moved this code to a separate function so that it is no
longer written out twice