Let actors be inside hollow surface regions
- enabled placement of actors inside hollow surface regions
- prevent parent/child regions from being neighbors if one of them is a
surface pointed in the opposite direction
- added check on child region that is an inner surface to ensure that
its parent is an outer surface that covers it fully
- improved membrane transitions to allow transition through cases when
region on other side of membrane was not a neighbor of the starting
region. Only works if region on other side of membrane is a neighbor of
the membrane region (i.e., not in nested grandparent cases)
- wrote code to potentially allow a 3D surface region that is a 3D shape
to have a 3D surface child. This is not enabled due to issues with
requiring transitions between regions that are not neighbors.