Corrected molecule reflections and absorption probability
- corrected molecule reflections during diffusion validation so that a
molecule is reflected off of the correct surface. Previously, it was
assumed that a molecule always reflects off of the outer boundary of its
starting region. Now a molecule can reflect off of whatever boundary it
could not successfully cross
- corrected equation for probability of absorption when a molecule
crosses an absorbing surface
- updated matlab import function to take an argument to indicate whether
to write the simulation output to a mat file