Bimolecular reactions with binding and unbinding radii
- added bimolecular chemical reactions to microscopic regime. User must
provide binding radius and (if applicable) unbinding radius. Potential
reactants are compared with other potential reactants in same region and
neighboring regions where the reaction could also take place. A molecule
cannot participate in more than one bimolecular reaction per time step.
Reaction site is found based on relative diffusion of reactants and must
be a valid straight line destination for both both reactants. If there
is a binding radius, trajectory of product molecules is also verified.
If 2 products, then they are placed along line between the 2 reactant
locations. If more than 2, then each is placed in a random direction.
- added sample configuration for a "crowding" scenario
- fixed index referencing of regions, actors, and reactions when
importing configuration files to Matlab
- Made warning more apparent when creating a large number of figures in
Matlab
- added indicator to check whether a particular global reaction can
occur in a given region
- added check on total number of chemical reactions in a surface region
before checking for surface reactions when a molecule crosses it. This
was needed because the region structure members to indicate surface
reactions are not created if a region has no reactions at all
- corrected bug where molecule is not correctly reflected off of a child
region surface