Corrected mesoscopic flow algorithm
- improve the mesoscopic flow algorithm by modifying the flow transition
rates. New method agrees with the first order solution of Fick's
diffusion equations at the interface. However, now have a special case
when the peclet value of the flow is greater than 2, in which case
diffusion is effectively ignored. The inaccuracy that results is
effectively a constraint on subvolume size when flows are present.